pyproject.toml
[project]
name = "blog-api"
version = "0.1.0"
description = "Simple Blogging Platform API"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "django>=4.2,<5.0",
    "psycopg2-binary>=2.9.0",
    "djangorestframework>=3.14.0",
    "python-dotenv>=1.0.0",
    "whitenoise>=6.5.0",
    "gunicorn>=21.0.0",
    "django-redis>=5.4.0",
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "django-debug-toolbar>=4.0.0",
    "django-extensions>=3.2.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "settings.dev"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing" 